Minimum Viable Project
Slack integration, improved logic, refactored code.
Everything is working nicely!

Just need to A) throttle requests to the server to avoid timeing out the
liquipedia api, B) deploy this app to the cloud somewhere., and C) make
a nice github readme page for the project.

Then this project will be complete.